+++ crates/holochain_cascade/src/authority/get_links_ops_query.rs
@@ -57,7 +57,6 @@ impl Query for GetLinksOpsQuery {
     type Output = Self::State;
 
     fn query(&self) -> String {
-        dbg!("query");
         let create = "
             SELECT Header.blob AS header_blob, DhtOp.type AS dht_type,
             DhtOp.validation_status AS status
@@ -87,7 +86,6 @@ impl Query for GetLinksOpsQuery {
             ));
         }
         if let Some(link_type) = &self.link_type {
-            dbg!(&link_type);
             common_query.push_str(&format!(
                 "
             AND