+++ crates/hdk/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - hdk: Adds external hash type for data that has a DHT location but does not exist on the DHT [#1298](https://github.com/holochain/holochain/pull/1298)
 - hdk: Adds compound hash type for linkable hashes [#1308](https://github.com/holochain/holochain/pull/1308)
 - hdk: Missing dependencies are fetched async for validation [#1268](https://github.com/holochain/holochain/pull/1268)
+- hdk: Possible to get links and link details filtered by link type [#1338](https://github.com/holochain/holochain/pull/1338/files)
 
 ## 0.0.127