+++ crates/holochain_integrity_types/CHANGELOG.md
@@ -4,4 +4,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Removed SerializedBytes derive from LinkType and LinkTag [#1338](https://github.com/holochain/holochain/pull/1338/files)
+
 ## 0.0.1